requirements.txt
pandas
vnstock
pyTelegramBotAPI
ta
python-dotenv
ipython
